+++ worldturtle/Graphics/WorldTurtle.hs
@@ -20,38 +20,38 @@ module Graphics.WorldTurtle
      -- * Running on a single turtle.
      -- $running
        runTurtle
+     , runTurtle'
      , TurtleCommand
      -- * Running a world of turtles.
      -- $multiturtle
      , runWorld
+     , runWorld'
      , WorldCommand
      , run 
      , (>/>)
      -- * Parallel animation
-     -- $parallel
-     , (<|>)
-     -- * Stop an animation
-     -- $empty
-     , empty
+     , (>!>)
      -- * Further documentation
      , module Graphics.WorldTurtle.Commands
      , module Graphics.WorldTurtle.Shapes
      , module Graphics.WorldTurtle.Color
      ) where
 
-import Control.Applicative (empty, (<|>))
+import Control.Monad.Parallel
 
 import Graphics.Gloss.Data.Display (Display (..))
 import qualified Graphics.Gloss.Data.ViewState as G
 import qualified Graphics.Gloss.Data.ViewPort as G
-import qualified Graphics.Gloss.Interface.Pure.Game as G
+import qualified Graphics.Gloss.Interface.IO.Game as G
 
 import Graphics.WorldTurtle.Color
 import Graphics.WorldTurtle.Commands
-import Graphics.WorldTurtle.Internal.Sequence (renderTurtle)
+import Graphics.WorldTurtle.Internal.Sequence (SequencePause, startSequence, resumeSequence, renderPause, defaultTSC)
 import Graphics.WorldTurtle.Internal.Commands ( TurtleCommand
-                                              , WorldCommand (..), seqW
-                                              , run)
+                                              , WorldCommand (..)
+                                              , run
+                                              , seqW
+                                              )
 import Graphics.WorldTurtle.Shapes
 
 -- | Takes a `TurtleCommand` and executes the command on an implicitly created
@@ -64,7 +64,49 @@ import Graphics.WorldTurtle.Shapes
 -- See also: `Graphics.WorldTurtle.Commands.makeTurtle`.
 runTurtle :: TurtleCommand () -- ^ Command sequence to execute.
           -> IO ()
-runTurtle c = runWorld $ makeTurtle >>= run c
+runTurtle = runTurtle' white
+
+-- | Variant of `runTurtle` which takes an additional background color parameter. 
+runTurtle' :: Color -- ^ Background color.
+          -> TurtleCommand () -- ^ Command sequence to execute.
+          -> IO ()
+runTurtle' bckCol c = runWorld' bckCol $ makeTurtle >>= run c
+
+-- | While `WorldCommand`s can be combined with `(>>)` to produce sequential
+--   instructions, we can also use the
+--   parallel animation operator `(>!>)` to achieve parallel instructions. 
+--   That is: animate two turtles at time!
+--
+--   Here is an example:
+--
+--   >  import Graphics.WorldTurtle
+--   >
+--   >  main :: IO ()
+--   >  main = runWorld $ do
+--   >    t1 <- makeTurtle' (0, 0) north green
+--   >    t2 <- makeTurtle' (0, 0) north red
+--   >
+--   >    -- Draw the anticlockwise and clockwise circles in sequence. 
+--   >    t1 >/> circle 90 >> t2 >/> circle (-90)
+--   >  
+--   >    clear
+--   >
+--   >    -- Draw the anticlockwise and clockwise circles in parallel.
+--   >    t1 >/> circle 90 >!> t2 >/> circle (-90)
+--
+--   Which would produce an animation like this
+--
+-- ![parallel and serial gif](docs/images/parallel_serial_turtles_2.gif)
+--
+-- Note that `(>!>)` is an alias for `bindM2`, and is defined as:
+-- 
+-- >  (>!>) = bindM2 (const . return)
+--
+(>!>) :: WorldCommand () -- ^ First command to execute in parallel
+      -> WorldCommand () -- ^ Second command to execute in parallel.
+      -> WorldCommand () -- ^ Result command
+(>!>) = bindM2 (const . return)
+infixl 3 >!>
 
 {- | `runWorld` takes a `WorldCommand` and produces the animation in a new
       window! 
@@ -91,23 +133,33 @@ runTurtle c = runWorld $ makeTurtle >>= run c
 -}
 runWorld :: WorldCommand () -- ^ Command sequence to execute
           -> IO ()
-runWorld tc = G.play display white 30 defaultWorld iterateRender input timePass
+runWorld = runWorld' white
+
+-- | Variant of `runWorld` which takes an additional background color parameter. 
+runWorld' :: Color -- ^ Background color
+          -> WorldCommand () -- ^ Command sequence to execute
+          -> IO ()
+runWorld' bckCol cmd = G.playIO display bckCol 30 (defaultWorld cmd) iterateRender input timePass
   where display = InWindow "World Turtle" (800, 600) (400, 300)
-        iterateRender w = G.applyViewPortToPicture 
-                               (G.viewStateViewPort $ state w)
-                        $ renderTurtle (seqW tc) (elapsedTime w)
+        iterateRender w = do
+           sq <- worldComputation w
+           let p = renderPause sq -- Render whatever is in the coroutine.
+           return $ G.applyViewPortToPicture (G.viewStateViewPort $ viewState w) p
         input e w 
              -- Reset key resets sim state (including unpausing). We 
              -- deliberately keep view state the same.
-             | isResetKey_ e = w { elapsedTime = 0, running = True }
+             | isResetKey_ e = return w {worldComputation = restartSequence cmd, running = True }
              -- Pause prevents any proceeding.
-             | isPauseKey_ e = w { running = not $ running w }
+             | isPauseKey_ e = return w { running = not $ running w }
              -- Let Gloss consume the command.
-             | otherwise = w { state = G.updateViewStateWithEvent e $ state w } 
+             | otherwise = return w { viewState = G.updateViewStateWithEvent e $ viewState w } 
         -- Increment simulation time if we are not paused.
         timePass f w
-         | running w = w { elapsedTime = f + elapsedTime w }
-         | otherwise = w
+         | running w = do
+               sq <- worldComputation w -- Grab previous sequence
+               sq' <- resumeSequence f sq -- Calculate new sequence
+               return w { worldComputation = return sq'}
+         | otherwise = return w
 
 -- | This is an infix version of `run` where the arguments are swapped.
 --
@@ -129,18 +181,21 @@ runWorld tc = G.play display white 30 defaultWorld iterateRender input timePass
 (>/>) = flip run
 infixl 4 >/>
 
-data World = World { elapsedTime :: !Float
-                   , running :: !Bool
-                   , state :: G.ViewState 
-                   }
+data World a = World { running :: !Bool
+                     , worldComputation:: IO (SequencePause a)
+                     , viewState :: G.ViewState
+                     }
 
-defaultWorld :: World
-defaultWorld = World 0 True 
-             $ G.viewStateInitWithConfig 
-             -- Easier to do this to have spacebar overwrite R.
-             $ reverse 
-             $ (G.CRestore, [(G.SpecialKey G.KeySpace, Nothing)])
-             : G.defaultCommandConfig
+restartSequence :: WorldCommand a -> IO (SequencePause a)
+restartSequence cmnd = startSequence defaultTSC (seqW cmnd)
+
+defaultWorld :: WorldCommand a -> World a
+defaultWorld cmd = World True (restartSequence cmd)
+                 $ G.viewStateInitWithConfig 
+                 -- Easier to do this to have spacebar overwrite R.
+                 $ reverse 
+                $ (G.CRestore, [(G.SpecialKey G.KeySpace, Nothing)])
+                : G.defaultCommandConfig
 
 -- | Tests to see if a key-event is the reset key.
 isResetKey_ :: G.Event -> Bool
@@ -159,11 +214,10 @@ isPauseKey_ _ = False
   To start animating a single turtle, you just pass your commands to
   `runTurtle` like so:
 
-  >    import Control.Monad (replicateM_)
   >    import Graphics.WorldTurtle
   >
   >    drawSquare :: Float -> TurtleCommand ()
-  >    drawSquare size = replicateM_ 4 $ forward size >> right 90
+  >    drawSquare size = repeatFor 4 $ forward size >> right 90
   >
   >    main :: IO ()
   >    main = runTurtle $ drawSquare 100
@@ -192,49 +246,4 @@ isPauseKey_ _ = False
    Notice that in a `WorldCommand` context we must create our own turtles with 
    `makeTurtle`! We them  apply the `TurtleCommand`
    on our turtles using the run operator `(>/>)`.
--}
-
-{- $parallel
-
-   #parallel#
-
-   While `WorldCommand`s can be combined with `(>>)` to produce sequential
-   instructions, we can also use the
-   alternative operator `(<|>)` to achieve parallel instructions. That is: 
-   animate two turtles at time!
-
-   Here is an example:
-
-   >  import Graphics.WorldTurtle
-   >
-   >  main :: IO ()
-   >  main = runWorld $ do
-   >    t1 <- makeTurtle' (0, 0) north green
-   >    t2 <- makeTurtle' (0, 0) north red
-   >
-   >    -- Draw the anticlockwise and clockwise circles in sequence. 
-   >    t1 >/> circle 90 >> t2 >/> circle (-90)
-   >  
-   >    clear
-   >
-   >    -- Draw the anticlockwise and clockwise circles in parallel.
-   >    t1 >/> circle 90 <|> t2 >/> circle (-90)
-
-   Which would produce an animation like this
-
-   ![parallel and serial gif](docs/images/parallel_serial_turtles_2.gif)
-
-   Note that the result of @x \<|\> y@ is:
-   
-   >>> x <|> y
-   x
-
-   when @x@ is not `Control.Applicative.empty`, otherwise the result is @y@.
--}
-
-{- $empty
-   If a `WorldCommand` is `Control.Applicative.empty`, then this stops this 
-   section of  animation and it does not progress. To this end 
-   `Control.Monad.guard` can be used to calculate when to stop part of an 
-   animation sequence.
 -}